requirements.txt
streamlit
openai
python-dotenv
PyMuPDF
Pillow
transformers
torch
sentence-transformers
opencv-python
matplotlib
easyocr